+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Changed
 
--   Make value fields nullable on `PageChooserBlock`, `SnippetChooserBlock`, `DocumentChooserBlock` and `ImageChooserBlock`
+-   Make value fields nullable on `PageChooserBlock`, `SnippetChooserBlock`, `DocumentChooserBlock` and `ImageChooserBlock` ([#396](https://github.com/torchbox/wagtail-grapple/pull/396))
 
 ## [0.25.1] - 2024-04-21